+++ app/src/organisms/Devices/RobotOverview.tsx
@@ -23,9 +23,9 @@ import { StyledText } from '../../atoms/text'
 import { useCurrentRunId } from '../ProtocolUpload/hooks'
 import { ChooseProtocolSlideout } from '../ChooseProtocolSlideout'
 import { Portal } from '../../App/portal'
-import { UpdateBuildroot } from '../../pages/Robots/RobotSettings/UpdateBuildroot'
 import { CONNECTABLE, UNREACHABLE } from '../../redux/discovery'
 import { UpdateRobotBanner } from '../UpdateRobotBanner'
+import { UpdateBuildroot } from './RobotSettings/UpdateBuildroot'
 import { RobotStatusBanner } from './RobotStatusBanner'
 import { RobotOverviewOverflowMenu } from './RobotOverviewOverflowMenu'
 import { useLights, useRobot, useIsRobotViewable } from './hooks'
