+++ app/src/organisms/Devices/RobotOverviewOverflowMenu.tsx
@@ -15,7 +15,6 @@ import {
 import { RUN_STATUS_RUNNING } from '@opentrons/api-client'
 import { checkShellUpdate } from '../../redux/shell'
 import { restartRobot } from '../../redux/robot-admin'
-import { UpdateBuildroot } from '../../pages/Robots/RobotSettings/UpdateBuildroot'
 import { UNREACHABLE } from '../../redux/discovery'
 import { home, ROBOT } from '../../redux/robot-controls'
 import { MenuItem } from '../../atoms/MenuList/MenuItem'
@@ -25,6 +24,7 @@ import { OverflowBtn } from '../../atoms/MenuList/OverflowBtn'
 import { Divider } from '../../atoms/structure'
 import { useCurrentRunStatus } from '../RunTimeControl/hooks'
 import { useMenuHandleClickOutside } from '../../atoms/MenuList/hooks'
+import { UpdateBuildroot } from './RobotSettings/UpdateBuildroot'
 
 import type { DiscoveredRobot } from '../../redux/discovery/types'
 import type { Dispatch, State } from '../../redux/types'
