+++ app/src/organisms/Devices/RobotSettings/UpdateBuildroot/__tests__/UpdateBuildroot.test.tsx
@@ -3,10 +3,10 @@ import * as React from 'react'
 import { mountWithStore, WrapperWithStore } from '@opentrons/components'
 import { mockConnectableRobot as mockRobot } from '../../../../../redux/discovery/__fixtures__'
 import * as Buildroot from '../../../../../redux/buildroot'
-import { UpdateBuildroot } from '..'
 import { VersionInfoModal } from '../VersionInfoModal'
 import { ViewUpdateModal } from '../ViewUpdateModal'
 import { InstallModal } from '../InstallModal'
+import { UpdateBuildroot } from '..'
 
 import type { State } from '../../../../../redux/types'
