+++ app/src/pages/Robots/RobotSettings/index.tsx
@@ -24,8 +24,8 @@ import { getRobotRestartRequired } from '../../../redux/robot-settings'
 import { SpinnerModalPage } from '@opentrons/components'
 import { CardContainer, CardRow } from '../../../atoms/layout'
 import { Page } from '../../../atoms/Page'
+import { UpdateBuildroot } from '../../../organisms/Devices/RobotSettings/UpdateBuildroot'
 import { ErrorModal } from '../../../molecules/modals'
-import { UpdateBuildroot } from './UpdateBuildroot'
 import { ConnectBanner } from './ConnectBanner'
 import { ReachableRobotBanner } from './ReachableRobotBanner'
 import { RestartRequiredBanner } from './RestartRequiredBanner'
